+++ src/App.tsx
@@ -1,11 +1,13 @@
 import { Scene } from './components/canvas/Scene'
 import { Upload } from './components/ui/Upload'
 import { RotationPanel } from './components/ui/RotationPanel'
+import { ResetAllPanel } from './components/ui/ResetAllPanel'
 
 function App() {
     return (
         <div className="w-full h-screen bg-gray-900 overflow-hidden relative">
             <Upload />
+            <ResetAllPanel />
             <RotationPanel />
             <Scene />
         </div>
@@ -13,3 +15,4 @@ function App() {
 }
 
 export default App
+